my_flask_app/requirements.txt
echo "gunicorn"
flask
flask-socketio
eventlet
gunicorn
requests
python-dotenv
openai
flask_sqlalchemy
flask_cors
